remove current_time_controller.rb
